Removed last errors that aren't just a function of Vienna dependence

+++ src/Util.hpp
@@ -16,7 +16,7 @@ using namespace std;
 class Util {
   public:
     static void baseReplace(string &base, const string& from, const string& to) {
-        int pos = 0;
+        unsigned long int pos = 0;
         while ((pos = base.find(from, pos)) != string::npos) {
             base.replace(pos, from.length(), to);
             pos += to.length();
